+++ README.md
@@ -1,9 +1,9 @@
-曾舒蕾 22009200443 提交
+曾舒蕾 22009200443 提交 第二题
 
 代码再model/
 
-演示的notebook为 model/run.ipynb
+演示的notebook为 model/timegan/run.ipynb
 
-启动脚本 run.sh
+启动脚本 model/timegan/run.sh
 
 报告为 timegan.pdf
