+++ model/timegan/data/stock.csv
