+++ model/timegan/main.py
