+++ model/timegan/metrics/general_rnn.py
