+++ model/timegan/models/timegan.py
