+++ model/timegan/output/test/train_time.pickle
