+++ model/timegan/run.sh
