+++ model/timegan/tensorboard/test/events.out.tfevents.1735908211.b25335db1e31.117989.0
