+++ model/timegan/tensorboard/test/events.out.tfevents.1735959079.b25335db1e31.131928.0
